+++ pkg/aws/v1alpha3/g8scontrolplane/validate_g8scontrolplane.go
@@ -54,6 +54,11 @@ func (v *Validator) ValidateCreate(request *admissionv1.AdmissionRequest) (bool,
 		return false, microerror.Maskf(parsingFailedError, "unable to parse awscontrol plane: %v", err)
 	}
 
+	err = aws.ValidateNamespace(&g8sControlPlane)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
 	err = aws.ValidateOrganizationLabelContainsExistingOrganization(context.Background(), v.k8sClient.CtrlClient(), &g8sControlPlane)
 	if err != nil {
 		return false, microerror.Mask(err)